otafilehosting/requirements.txt
flask
cryptography
python-jose